Fix: ViewModel creation and getAlbumsFromAPI work
Need to check if FlickrDataProvider.isContentOnDB() is working correctly or not. First time it's called it returns true, but the DB should be empty in that moment. After it's called, the searchAlbumsInDataBase() method returns null, so there is something wrong in the isContentOnDB() method. Next step is to continue with PhotoFragment. Didn't need to add the following lines on the build.gladle script:

implementation 'androidx.lifecycle:lifecycle-process:2.2.0'
implementation 'androidx.lifecycle:lifecycle-service:2.2.0'
implementation 'androidx.lifecycle:lifecycle-viewmodel-savedstate:2.2.0'
annotationProcessor 'androidx.lifecycle:lifecycle-compiler:2.2.0'
implementation 'androidx.fragment:fragment:1.2.2'