Feature: ActionBar, Preferences (order), Receiver
The action bar works, so do the preferences that sort the recycler views (albums and photos). A broadcast receiver was implemented to see if there is connection or not. The snackbar works ok. The method that checks for content on the DB seems not to work without internet connection, wich needs to be solved